requirements.txt
google-generativeai>=0.3.0
colorama>=0.4.0
pyinstaller>=5.0  # For building only